requirements.txt
akracer==0.0.14
akshare==1.18.35
baostock>=0.8.8
efinance>=0.5.5
anyio==4.12.1
beautifulsoup4==4.14.3
certifi==2026.2.25
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
curl_cffi==0.13.0
decorator==5.2.1
et_xmlfile==2.0.0
frozendict==2.4.7
h11==0.16.0
html5lib==1.1
httpcore==1.0.9
httpx==0.25.2
idna==3.11
jsonpath==0.82.2
lxml==6.0.2
mootdx==0.11.7
multitasking==0.0.12
# pandas-ta 依赖 numba；numba 0.61.x 要求 numpy<2.3，与 numpy 2.4.x 冲突，故锁定 2.2.6（满足 pandas-ta>=2.2.6）
numpy==2.2.6
openpyxl==3.1.5
pandas==3.0.1
pandas-ta==0.4.71b0
peewee==4.0.0
platformdirs==4.9.2
prettytable==3.17.0
protobuf==6.33.5
py-mini-racer==0.6.0
pycparser==3.0
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
six==1.17.0
sniffio==1.3.1
soupsieve==2.8.3
tabulate==0.10.0
tdxpy==0.2.7
tenacity==8.5.0
tqdm==4.67.3
typing_extensions==4.15.0
urllib3==2.6.3
wcwidth==0.6.0
webencodings==0.5.1
websockets==16.0
xlrd==2.0.2
yfinance==1.2.0
pyyaml>=6.0
python-dotenv>=1.0.0
# llm_enhancer（OpenAI 兼容接口，如 DeepSeek）
openai>=1.40.0
